+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Simplified fetch urls to relatives
 - Improved code for passed forecast data error handling in utils.py
 - Added new tests for forecast longer than 24h by changing parameter `delta_forecast`
+- Added new files for updated PV modules and inverters using PVLib
 ### Fix
 - style.css auto format and adjusted table styling